% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{AHP}
\alias{AHP}
\title{AHP: Analytic Hierarchy Process}
\usage{
AHP(A)
}
\arguments{
\item{A}{a numeric matrix, i.e. pairwise comparison matrix}
}
\value{
a list object that contains: w (Weight vector), CR (Consistency ratio),
Lmax (Maximum eigenvalue), CI (Consistency index)
}
\description{
AHP is a multi-criteria decision
analysis method developed by Saaty, which can also be used to
determine indicator weights.
}
\examples{
A = matrix(c(1,   1/2, 4, 3,   3,
             2,   1,   7, 5,   5,
             1/4, 1/7, 1, 1/2, 1/3,
             1/3, 1/5, 2, 1,   1,
             1/3, 1/5, 3, 1,   1), byrow = TRUE, nrow = 5)
AHP(A)
}
