% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{.ts_theme}
\alias{.ts_theme}
\title{Shared ggplot2 theme for mathmodels time series plots}
\usage{
.ts_theme(base_size = 11)
}
\description{
Shared ggplot2 theme for mathmodels time series plots
}
\keyword{internal}
