% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{.validate_ts_input}
\alias{.validate_ts_input}
\title{Validate and coerce input to ts object}
\usage{
.validate_ts_input(x, frequency = NULL)
}
\description{
Validate and coerce input to ts object
}
\keyword{internal}
