% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{drop_na_ts_df}
\alias{drop_na_ts_df}
\title{Drop Missing Values from a \code{ts_df}}
\usage{
drop_na_ts_df(x)
}
\arguments{
\item{x}{A \code{ts_df}.}
}
\value{
A \code{ts_df} with rows whose \code{value} is \code{NA} removed.
}
\description{
Drop Missing Values from a \code{ts_df}
}
\examples{
x = ts_df(data.frame(time = 1:4, value = c(10, NA, 12, 13)))
drop_na_ts_df(x)
}
