% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{is_ts_df}
\alias{is_ts_df}
\title{Check Whether an Object Is a \code{ts_df}}
\usage{
is_ts_df(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
Check Whether an Object Is a \code{ts_df}
}
\examples{
x = ts_df(data.frame(time = 1:3, value = c(10, 12, 15)))
is_ts_df(x)
}
