% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmapwidget.R
\name{logmapwidgetOutput}
\alias{logmapwidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
logmapwidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

