% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_plot.R
\name{plot_compartments}
\alias{plot_compartments}
\title{Plot compartment trajectories}
\usage{
plot_compartments(df, compartments = NULL)
}
\arguments{
\item{df}{A data frame returned by \code{ode_solver()} or any
built-in model function.}

\item{compartments}{Character vector of compartment names to plot.  When
\code{NULL} (the default) every column except \code{time} is plotted.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Draws one line per compartment over time.  Use \code{compartments} to
select a subset of the state variables; the default is to plot all
columns except \code{time}.
}
\examples{
sir = model_sir(
  init   = c(S = 990, I = 10),
  params = c(beta = 0.002, gamma = 0.1),
  times  = seq(0, 50, by = 0.1)
)
plot_compartments(sir)
plot_compartments(sir, compartments = c("S", "I"))

}
