% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_plot.R
\name{plot_incidence}
\alias{plot_incidence}
\title{Plot incidence curves}
\usage{
plot_incidence(df)
}
\arguments{
\item{df}{A data frame returned by \code{ode_solver()} or any
built-in model function.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Calls \code{compute_incidence()} and draws one line for each
incidence metric (new infections from S, new infections from I,
new deaths).
}
\examples{
sir = model_sir(
  init   = c(S = 990, I = 10),
  params = c(beta = 0.002, gamma = 0.1),
  times  = seq(0, 50, by = 0.1)
)
plot_incidence(sir)

}
