% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidemic_plot.R
\name{plot_infectious_curve}
\alias{plot_infectious_curve}
\title{Plot infectious population}
\usage{
plot_infectious_curve(df)
}
\arguments{
\item{df}{A data frame with columns \code{time} and \code{I}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Draws a single line showing the number of infectious individuals
over time.  Requires a column named \code{"I"}.
}
\examples{
sir = model_sir(
  init   = c(S = 990, I = 10),
  params = c(beta = 0.002, gamma = 0.1),
  times  = seq(0, 50, by = 0.1)
)
plot_infectious_curve(sir)

}
