% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot a Time Series}
\usage{
plot_ts(x, title = "Time Series", x_lab = "Time", y_lab = "Value")
}
\arguments{
\item{x}{A \code{ts_df}.}

\item{title}{Plot title.}

\item{x_lab, y_lab}{Axis labels.}
}
\value{
A ggplot object.
}
\description{
Plot a Time Series
}
\examples{
plot_ts(as_ts_df(AirPassengers))
}
