% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_forecast}
\alias{plot_ts_forecast}
\title{Plot Forecasts}
\usage{
plot_ts_forecast(
  x,
  forecast_tbl,
  level = c(80, 95),
  by = NULL,
  title = "Forecast"
)
}
\arguments{
\item{x}{Original \code{ts_df}.}

\item{forecast_tbl}{Result from \code{ts_forecast()}.}

\item{level}{Confidence levels to draw when present.}

\item{by}{Step size for future time values.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot Forecasts
}
\examples{
fit = ts_ets(as_ts_df(log(AirPassengers)))
fc = ts_forecast(fit, h = 3)
plot_ts_forecast(as_ts_df(log(AirPassengers)), fc, by = 1)
}
