% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_garch}
\alias{plot_ts_garch}
\title{GARCH Volatility Plot}
\usage{
plot_ts_garch(garch_result, title = "GARCH Volatility Analysis")
}
\arguments{
\item{garch_result}{Result list from \code{ts_garch()}.}

\item{title}{Character. Plot title.}
}
\value{
A \code{patchwork} composite \code{ggplot}.
}
\description{
Four-panel overview: (1) observed returns with \eqn{\pm 2\sigma} bands, (2) conditional
variance, (3) standardised residuals, (4) histogram of standardised residuals.
}
\examples{
r = diff(log(AirPassengers))
res = ts_garch(r)
plot_ts_garch(res)

}
