% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_pacf}
\alias{plot_ts_pacf}
\title{Partial Autocorrelation Plot}
\usage{
plot_ts_pacf(x, max_lag = 40L, title = "PACF")
}
\arguments{
\item{x}{A complete \code{ts_df} with no missing values.}

\item{max_lag}{Maximum lag.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Partial Autocorrelation Plot
}
\examples{
plot_ts_pacf(as_ts_df(log(AirPassengers)))
}
