% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_residuals}
\alias{plot_ts_residuals}
\title{Residual Diagnostic Plot}
\usage{
plot_ts_residuals(model_result, max_lag = 30L, title = "Residual Diagnostics")
}
\arguments{
\item{model_result}{Result from \code{ts_ets()} or \code{ts_sarima()}.}

\item{max_lag}{Maximum lag.}

\item{title}{Plot title.}
}
\value{
A patchwork plot.
}
\description{
Residual Diagnostic Plot
}
\examples{
fit = ts_sarima(as_ts_df(log(AirPassengers)), stepwise = TRUE, approximation = TRUE)
plot_ts_residuals(fit)
}
