% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_stl}
\alias{plot_ts_stl}
\title{STL Decomposition Plot}
\usage{
plot_ts_stl(stl_result, title = "STL Decomposition")
}
\arguments{
\item{stl_result}{Result from \code{ts_stl()}.}

\item{title}{Plot title.}
}
\value{
A patchwork plot.
}
\description{
STL Decomposition Plot
}
\examples{
plot_ts_stl(ts_stl(as_ts_df(AirPassengers)))
}
