% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predpreywidget.R
\name{predpreywidgetOutput}
\alias{predpreywidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
predpreywidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

