% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_reg.R
\name{reg_plot_residuals}
\alias{reg_plot_residuals}
\title{Residual Diagnostics}
\usage{
reg_plot_residuals(model_result)
}
\arguments{
\item{model_result}{A list returned by any \code{reg_*()} function.}
}
\value{
A ggplot-based composite.
}
\description{
Returns a multi-panel diagnostic plot from a fitted regression model.
}
