% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competitionwidget.R
\name{renderCompetitionwidget}
\alias{renderCompetitionwidget}
\title{Widget render function for use in Shiny}
\usage{
renderCompetitionwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

