% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hysteresiswidget.R
\name{renderHysteresiswidget}
\alias{renderHysteresiswidget}
\title{Widget render function for use in Shiny}
\usage{
renderHysteresiswidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

