% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticwidget.R
\name{renderLogisticwidget}
\alias{renderLogisticwidget}
\title{Widget render function for use in Shiny}
\usage{
renderLogisticwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

