% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmapwidget.R
\name{renderLogmapwidget}
\alias{renderLogmapwidget}
\title{Widget render function for use in Shiny}
\usage{
renderLogmapwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

