% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predpreywidget.R
\name{renderPredpreywidget}
\alias{renderPredpreywidget}
\title{Widget render function for use in Shiny}
\usage{
renderPredpreywidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

