% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirwidget.R
\name{renderSirwidget}
\alias{renderSirwidget}
\title{Widget render function for use in Shiny}
\usage{
renderSirwidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

