% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirwidget.R
\name{sirwidgetOutput}
\alias{sirwidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
sirwidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

