% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{ts_ets}
\alias{ts_ets}
\title{ETS Model Fitting}
\usage{
ts_ets(x, model = "ZZZ", ...)
}
\arguments{
\item{x}{A complete \code{ts_df} with no missing values.}

\item{model}{ETS model string passed to \code{forecast::ets()}.}

\item{...}{Additional arguments passed to \code{forecast::ets()}.}
}
\value{
A list containing tidy model summaries, the raw model, and input.
}
\description{
ETS Model Fitting
}
\examples{
ts_ets(as_ts_df(log(AirPassengers)))
}
