% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{ts_forecast}
\alias{ts_forecast}
\title{Generate Forecasts}
\usage{
ts_forecast(model_result, h = 12, level = c(80, 95))
}
\arguments{
\item{model_result}{Result from \code{ts_ets()} or \code{ts_sarima()}.}

\item{h}{Forecast horizon.}

\item{level}{Confidence levels.}
}
\value{
A tibble with point forecasts and intervals.
}
\description{
Generate Forecasts
}
\examples{
fit = ts_ets(as_ts_df(log(AirPassengers)))
ts_forecast(fit, h = 3)
}
