% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{ts_sarima}
\alias{ts_sarima}
\title{SARIMA Model Fitting}
\usage{
ts_sarima(
  x,
  order = NULL,
  seasonal = NULL,
  stepwise = TRUE,
  approximation = TRUE,
  ...
)
}
\arguments{
\item{x}{A complete \code{ts_df} with no missing values.}

\item{order}{Optional non-seasonal ARIMA order.}

\item{seasonal}{Optional seasonal specification.}

\item{stepwise, approximation}{Passed to \code{forecast::auto.arima()}.}

\item{...}{Additional arguments passed to the forecast fitting function.}
}
\value{
A list containing tidy model summaries, the raw model, and input.
}
\description{
SARIMA Model Fitting
}
\examples{
ts_sarima(as_ts_df(log(AirPassengers)), stepwise = TRUE, approximation = TRUE)
}
