% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{ts_test}
\alias{ts_test}
\title{Stationarity Tests for a Time Series}
\usage{
ts_test(x, adf_lags = NULL)
}
\arguments{
\item{x}{A complete \code{ts_df} with no missing values.}

\item{adf_lags}{Optional lag for the ADF test.}
}
\value{
A tibble with test results.
}
\description{
Runs ADF and KPSS tests on a complete \code{ts_df}.
}
\examples{
ts_test(as_ts_df(log(AirPassengers)))
}
