% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{validate_ts_df}
\alias{validate_ts_df}
\title{Validate a \code{ts_df}}
\usage{
validate_ts_df(x, require_complete = TRUE, require_no_na = FALSE)
}
\arguments{
\item{x}{A \code{ts_df} object.}

\item{require_complete}{If \code{TRUE}, first and last values must not be \code{NA}.}

\item{require_no_na}{If \code{TRUE}, no values may be \code{NA}.}
}
\value{
A sorted \code{ts_df} object.
}
\description{
Validate a \code{ts_df}
}
\examples{
x = ts_df(data.frame(time = 1:3, value = c(10, 12, 15)))
validate_ts_df(x)
}
