% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHP.R
\name{AHP}
\alias{AHP}
\title{FUNCTION_TITLE}
\usage{
AHP(A)
}
\arguments{
\item{A}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
