% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_ts.R
\name{plot_ts_stl}
\alias{plot_ts_stl}
\title{Plot STL Decomposition Components}
\usage{
plot_ts_stl(stl_result, title = "STL Decomposition")
}
\arguments{
\item{stl_result}{Result list from \code{ts_stl()}.}

\item{title}{Character. Plot title.}
}
\value{
A \code{patchwork} composite \code{ggplot}.
}
\description{
Stacked four-panel plot: observed, trend, seasonal, remainder.
}
\examples{
data(AirPassengers)
res = ts_stl(AirPassengers)
plot_ts_stl(res)

}
